Fix: PlayerCardCalculation 제거 && Player 가 스스로 자기의 손패를 계산하도록 변경
